COG-532 refactoring: extracted interface for cognito facade, interface introduced improves code readability. Package-private visibility for facade implementation hides details of implementation for other packages.
