+++ src/test/java/gov/ca/cwds/Constants.java
@@ -23,4 +23,18 @@ public final class Constants {
   public static final String TOKEN_STORE_URL = DATABASE_URL + SCHEMA_EQUALS + TOKEN_STORE_SCHEMA;
   public static final String CMS_STORE_URL = DATABASE_URL + SCHEMA_EQUALS + CMS_STORE_SCHEMA;
 
+  public static final String USERPOOL = "userpool";
+  public static final String USER_NO_RACFID_ID = "2be3221f-8c2f-4386-8a95-a68f0282efb0";
+  public static final String USER_WITH_RACFID_ID = "24051d54-9321-4dd2-a92f-6425d6c455be";
+  public static final String ERROR_USER_ID = "errorUserId";
+  public static final String ABSENT_USER_ID = "absentUserId";
+  public static final String USER_WITH_RACFID_AND_DB_DATA_ID =
+      "d740ec1d-80ae-4d84-a8c4-9bed7a942f5b";
+  public static final String USER_WITH_NO_PHONE_EXTENSION = "d740ec1d-66ae-4d84-a8c4-8bed7a942f5b";
+  public static final String NEW_USER_SUCCESS_ID = "17067e4e-270f-4623-b86c-b4d4fa527a34";
+  public static final String NEW_USER_ES_FAIL_ID = "08e14c57-6e5e-48dd-8172-e8949c2a7f76";
+  public static final String ES_ERROR_CREATE_USER_EMAIL = "es.error@create.com";
+  public static final String SOME_PAGINATION_TOKEN = "somePaginationToken";
+
+
 }
